### 2022-02-10 #4144 @whedon
[PRE REVIEW]: TrackerControl: Transparency and Choice around App Tracking
**Submitting author:** @kasnder (<a href="http://orcid.org/0000-0002-7412-8731">Konrad Kollnig</a>)
**Repository:** https://github.com/TrackerControl/tracker-control-android
**Branch with paper.md** (empty if default branch):
**Version:**
