### 2022-02-17 e676a8c
New translations short_description.txt (Basque)
