+++ fastlane/metadata/android/eu-ES/short_description.txt
@@ -1,1 +1,1 @@
-TrackerControl allows to monitor and control hidden data collection in apps.
+TrackerControl-ek app-en ezkutuko informazio bilketa kontrolatzea baimentzen du.
