### 2022-02-18 d41178f
Clean up the state adapter a little
