### 2022-02-18 809f07a
Rename DetailsPagesAdapter to DetailsStateAdapter
