+++ app/src/main/java/net/kollnig/missioncontrol/DetailsStateAdapter.java
@@ -34,21 +34,21 @@ import net.kollnig.missioncontrol.details.TrackersFragment;
  * A [FragmentStateAdapter] that returns a fragment corresponding to
  * one of the sections/tabs/pages.
  */
-public class DetailsPagesAdapter extends FragmentStateAdapter {
+public class DetailsStateAdapter extends FragmentStateAdapter {
     @StringRes
     private static int[] TAB_TITLES = {
             R.string.tab_trackers,
             R.string.tab_actions,
             R.string.tab_countries,
     };
-    private final String TAG = DetailsPagesAdapter.class.getSimpleName();
+    private final String TAG = DetailsStateAdapter.class.getSimpleName();
     private final int mUid;
 
     private final TrackersFragment fTrackers;
     private final ActionsFragment fActions;
     private CountriesFragment fCountries;
 
-    public DetailsPagesAdapter(FragmentActivity fa, String appId, String appName, int uid) {
+    public DetailsStateAdapter(FragmentActivity fa, String appId, String appName, int uid) {
         super(fa);
 
         mUid = uid;
