### 2022-03-07 #4144 @sbenthall
@gradvohl Thanks for your question. I am in the process of searching for a second reviewer. When I have one confirmed, I will begin the review process, which will create a new issue with a checklist for reviewers.

If you would like to do your review ahead of time, I suggest you use this checklist:

https://joss.readthedocs.io/en/latest/review_checklist.html

... and record your review responses in a way that makes it easy to transfer them into the new form when that issue is made.
