### 2022-03-22 e4c2e0d
Update Disconnect and add new Email category
