+++ app/src/main/assets/ip_blocklist.txt
@@ -423,8 +423,6 @@
 144.76.37.44
 144.76.39.37
 146.101.66.152
-146.185.22.76
-146.185.22.80
 146.82.109.225
 146.82.138.194
 146.82.202.112
@@ -442,7 +440,6 @@
 147.136.250.60
 147.202.86.254
 147.208.175.16
-149.13.34.195
 149.210.134.160
 151.1.244.150
 151.1.244.178
@@ -1009,7 +1006,6 @@
 174.36.240.233
 174.36.45.48
 174.36.58.233
-174.36.6.58
 174.37.147.184
 174.37.154.104
 174.37.154.236
@@ -1641,8 +1637,6 @@
 195.189.143.135
 195.189.236.35
 195.193.95.230
-195.198.236.113
-195.198.236.63
 195.2.64.1
 195.2.82.69
 195.20.224.83
@@ -1650,9 +1644,7 @@
 195.20.249.19
 195.204.232.34
 195.204.232.40
-195.210.87.1
 195.210.87.2
-195.210.91.107
 195.210.91.77
 195.210.96.10
 195.210.96.8
@@ -1845,10 +1837,8 @@
 198.74.57.251
 198.74.62.254
 198.77.36.41
-198.78.206.126
 198.78.210.126
 198.78.212.126
-198.78.218.126
 198.78.219.126
 198.80.189.106
 198.87.214.21
@@ -1946,7 +1936,6 @@
 199.93.54.126
 199.93.57.126
 199.93.57.254
-199.93.62.126
 199.95.206.210
 2.16.11.234
 2.16.139.234
@@ -2088,7 +2077,6 @@
 204.11.52.68
 204.11.56.21
 204.12.68.104
-204.12.71.5
 204.12.89.69
 204.12.89.76
 204.128.175.252
@@ -2122,16 +2110,11 @@
 204.155.175.242
 204.16.173.2
 204.16.252.98
-204.160.107.126
 204.160.120.125
 204.160.98.126
 204.160.98.253
 204.176.177.140
 204.176.88.14
-204.179.240.184
-204.179.240.226
-204.179.240.43
-204.179.240.79
 204.181.14.4
 204.183.113.68
 204.197.240.133
@@ -2159,8 +2142,6 @@
 204.253.104.220
 204.71.191.220
 204.71.191.94
-204.78.32.60
-204.78.50.226
 204.79.197.200
 204.8.169.150
 204.8.169.182
@@ -2171,11 +2152,8 @@
 204.9.177.90
 204.9.178.17
 204.93.155.71
-205.128.66.126
-205.128.67.254
 205.128.68.126
 205.128.78.126
-205.128.79.126
 205.128.82.254
 205.128.85.126
 205.128.90.126
@@ -2296,8 +2274,6 @@
 206.196.125.120
 206.200.251.66
 206.204.187.42
-206.207.85.33
-206.207.87.8
 206.214.223.6
 206.214.223.81
 206.222.17.10
@@ -2307,9 +2283,7 @@
 206.223.183.156
 206.252.153.4
 206.252.157.58
-206.33.34.126
 206.33.54.126
-206.33.55.126
 206.47.74.31
 206.51.226.142
 206.55.125.138
@@ -2326,11 +2300,8 @@
 207.123.33.124
 207.123.33.126
 207.123.34.126
-207.123.39.254
 207.123.40.126
-207.123.40.254
 207.123.57.126
-207.123.61.126
 207.13.199.239
 207.150.160.71
 207.150.160.72
@@ -2391,9 +2362,6 @@
 207.171.175.96
[…]
