### 2022-03-24 #4270 @editorialbot
[REVIEW]: TrackerControl: Transparency and Choice around App Tracking
**Submitting author:** @kasnder (<a href="http://orcid.org/0000-0002-7412-8731">Konrad Kollnig</a>)
**Repository:** https://github.com/TrackerControl/tracker-control-android
**Branch with paper.md** (empty if default branch): master
**Version:** 2022061501
