### 2022-03-25 #4270 @kasnder
Hi @sbenthall

Thank you for getting the processed started. The paper is in the branch `joss`.
