### 2022-03-28 #4270 @gradvohl
# Reviewer report for JOSS paper #4270
**Paper title**: TrackerControl: Transparency and Choice around App Tracking

## Review criteria
**Note**: whenever an item is marked with :warning: or :no_entry:, you can get more information about that evaluation in the [**Reviewer Comments**](#reviewer-comments) section).

* A list of the authors of the software: :heavy_check_mark:
* Author affiliations: :heavy_check_mark:
* A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience: :heavy_check_mark:
* A clear statement of need that illustrates the purpose of the software: :heavy_check_mark:
* Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it: [:warning:](#Mentions-any-ongoing-research-projects)
* A list of key references, including a link to the software archive:  [:warning:](#a-list-of-key-references)

## Review items
* Software license: :thumbsup: (good -- GPL-3.0 License)
* Substantial scholarly effort: :thumbsup: (good -- Please see the [**Scholarly effort**](#scholarly-effort)  section in the [**Reviewer Comments**](#reviewer-comments) section)
* Documentation:
    * A statement of need: :thumbsup: (**good**)
    * Installation instructions: :thumbsup: (**good**)
    * Example usage: :thumbsdown: (**bad** -- Please see the [**Example usage**](#example-usage)  section in the [**Reviewer Comments**](#reviewer-comments) section)
    * API documentation: :thumbsdown: (**bad** -- Please see the [**API Docs**](#api-docs)  section in the [**Reviewer Comments**](#reviewer-comments) section)
    * Community guidelines:  :thumbsdown: (**bad** -- Please see the [**Community guidelines**](#community-guidelines)  section in the [**Reviewer Comments**](#reviewer-comments) section)
    * Functionality: :thumbsup: (**good**)
    * Tests: [:warning:](#tests-on-source-code)

* Software paper
	* The ``paper.pdf`` is not on the ``master`` branch [:warning:](#about-paper.md-file).
    * Authors: :thumbsup: (**good**)
    * A statement of need: :thumbsup: (**good**)
    * Quality of writing: :thumbsup: (**good**)
    * References: :thumbsup: (**good**)

## Test environment
The app was downloaded and tested on an Android smartphone.

## Reviewer comments
The following comments refer to items I considered the author could improve to meet JOSS submission requirements.

### About ``paper.md`` file
Although the ``paper.md`` file is in the ``joss`` branch, it will be better if it is in the ``master`` branch, especially when the authors release the app's final version.

### Mentions any ongoing research projects
I did not find any mention about ongoing research projects. It is not a requirement for JOSS, but if the authors are aware of any project that uses the TrackControl app, I recommend citing that project or, perhaps, writing about its potential in other research projects.

### A list of key references
There is a list of key references on the ``paper.md`` file. However, there are no references in the repository. I do not think that list is mandatory in the repository, but it is worth checking.

### Example usage
I did not find any tutorial or examples for using the software in any file on the repository. Therefore, I suggest that the author include a ``docs`` directory with a tutorial or screenshots describing the software's typical usage (step-by-step).

### Scholarly effort
The software had a substantial effort to build it. However, I consider that the software is much more commercial than academic software. I highlight that my opinion is not a demerit of the software.

### API Docs
I did not find any documentation about the code regarding the API documentation, neither in the source code nor in a ``docs`` directory (the repository does not have a ``docs`` directory). Therefore, I strongly recommend the documentation of the source code using a standard.
My suggestion is to use the Javadocs or [Doxygen](https://www.doxygen.nl/index.html) software. Doxygen can generate the docs (in HTML or PDF) from the documented source files.

### Community guidelines
There are indications of communities in different media (Telegram, Matrix etc). However, there are no community guidelines. My suggestion is to create a **Contributing** section (or a ``CONTRIBUTING.md`` file) with more specific instructions about how to contribute.

### Tests on source code
We ran the application tests on an Android  smartphone. It works nicely. However, we did not find any automatic tests on the source code or instructions about how to do the tests. It is not mandatory to specify those instructions, but it would be nice if the authors provided instructions on how to compile and test the source code.

### Other Suggestions
Please, consider including the ``CITATION.cff`` file with information about how to cite the ``TrackerControl`` software. In addition, information about software citation in GitHub is available on [GitHub Docs](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files).

## Conclusion
Analyzing all the aforementioned aspects, I rank the paper as **major revisions** needed. Although most of those revisions are very easy to implement, I consider the lack of documentation and the absence of the ``paper.md`` file in the master branch prevents acceptance of the paper at this time.
