### 2022-03-30 1458355
New translations full_description.txt (Japanese)
