+++ fastlane/metadata/android/ja-JP/full_description.txt
@@ -1,13 +1,13 @@
-TrackerControl allows users to monitor and control the widespread, ongoing, hidden data collection in mobile apps about user behaviour ('tracking').
+TrackerControlを使用すると、普段気付かないような、広範囲にわたる、データ収集(トラッキング)を監視・制御できます。
 
-To detect tracking, TrackerControl combines the power of the Disconnect blocklist, used by Firefox, and our in-house blocklist is used, created from analysing <i>~2 000 000</i> apps! Additionally, TrackerControl supports custom blocklists.
+TrackerControl は、トラッキングを検出するため、Firefox で使用されている Disconnect blocklist と、 <i>2,000</i> 以上のアプリを解析することによって作成された、独自のブロックリストを組み合せて使用します。 さらに、TrackerControl はカスタムブロックリストをサポートしています。
 
 This approach
-- reveals the companies behind tracking,
-- allows to block tracking selectively, and
-- exposes the purposes of tracking, such as analytics or advertising.
+- トラッキングの背後にある企業を明らかにし、
+- 選択的にトラッキングをブロックすることができます。
+- また、分析や広告など、トラッキングの目的も明らかにします。
 
-The app also aims to educate about your rights under Data Protection Law, such the EU General Data Protection Regulation (GDPR).
+他にも、EU一般データ保護規則(GDPR)など、データ保護法に基づくあなたのの権利についても知ることができます。
 
 Under the hood, TrackerControl uses Android's VPN functionality, to analyse apps' network communications <i>locally on the Android device</i>. This is accomplished through a local VPN server, to enable network traffic analysis by TrackerControl.
 
