### 2022-03-30 3ee51a0
New translations full_description.txt (Japanese)
