### 2022-03-30 0347ba4
New translations short_description.txt (Japanese)
