+++ fastlane/metadata/android/ja-JP/short_description.txt
@@ -1,1 +1,1 @@
-TrackerControl allows to monitor and control hidden data collection in apps.
+TrackerControl は、アプリの隠されたデータトラッキングを監視・制御することができます。
