### 2022-03-30 14cb150
New translations full_description.txt (Japanese)
