+++ fastlane/metadata/android/ja-JP/full_description.txt
@@ -13,5 +13,5 @@ TrackerControl は、Android の VPN機能 を使用して、アプリのネッ�
 
 ルート権限は不要です。 データの安全性を確保するため、外部のVPNサーバーは使用していません！ ウェブサイトやアプリにトラッカーを隠ためによく使われるテクニックである、 <i>DNSクローキング</i> からもあなたを守ります。
 
-TrackerControlは、研究目的のプロジェクトであるため、これまでも、これからも、常にフリーで、オープンソースです。
+TrackerControl は、研究目的のプロジェクトなため、これまでも、これからも、常にフリーで、オープンソースです。
 
