+++ paper.bib
@@ -1,0 +1,208 @@
+@misc{netguard,
+  author = {Marcel Bokhorst},
+  title = {NetGuard: A simple way to block access to the internet per app},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/M66B/NetGuard}
+}
+
+@misc{izzy,
+  title        = {TrackerControl},
+  author       = {Oxford HCC},
+  year         = 2021,
+  url          = {https://apt.izzysoft.de/fdroid/index/apk/net.kollnig.missioncontrol},
+  howpublished = {\url{https://apt.izzysoft.de/fdroid/index/apk/net.kollnig.missioncontrol}}
+}
+
+@inproceedings{enck_taintdroid_2010,
+  title = {{{TaintDroid}}: {{An Information}}-Flow {{Tracking System}} for {{Realtime Privacy Monitoring}} on {{Smartphones}}},
+  booktitle = {Proceedings of the 9th {{USENIX Conference}} on {{Operating Systems Design}} and {{Implementation}}},
+  author = {Enck, William and Gilbert, Peter and Chun, Byung-Gon and Cox, Landon P. and Jung, Jaeyeon and McDaniel, Patrick and Sheth, Anmol N.},
+  year = {2010},
+  pages = {393--407},
+  series = {{{OSDI}}'10}
+}
+
+@InProceedings{agarwal_protectmyprivacy_2013,
+  title = {{{ProtectMyPrivacy}}: Detecting and Mitigating Privacy Leaks on {{iOS}} Devices Using Crowdsourcing},
+  shorttitle = {{{ProtectMyPrivacy}}},
+  booktitle = {Proceeding of the 11th Annual International Conference on {{Mobile}} Systems, Applications, and Services - {{MobiSys}} '13},
+  author = {Yuvraj Agarwal and Malcolm Hall},
+  year = {2013},
+  pages = {97},
+  publisher = {{ACM Press}},
+  address = {{Taipei, Taiwan}},
+  doi = {10.1145/2462456.2464460},
+  isbn = {978-1-4503-1672-9},
+  language = {en},
+}
+
+@inproceedings{privacyguard_vpn_2015,
+  author = {Yihang Song and Urs Hengartner},
+  title = {{{PrivacyGuard}}: A {{VPN}}-based Platform to Detect Information Leakage on Android Devices},
+  booktitle = {Proceedings of the 5th Annual ACM CCS Workshop on Security and Privacy in Smartphones and Mobile Devices},
+  series = {SPSM '15},
+  year = {2015},
+  pages = {15--26},
+  doi = {10.1145/2808117.2808120},
+}
+
+@inproceedings{nomoads_2018,
+  author = {Anastasia Shuba and Athina Markopoulou and Zubair Shafiq},
+  year = {2018},
+  month = {10},
+  pages = {125-140},
+  title = {NoMoAds: Effective and Efficient Cross-App Mobile Ad-Blocking},
+  booktitle = {Proceedings on Privacy Enhancing Technologies 2018},
+  doi = {10.1515/popets-2018-0035},
+}
+
+@Article{reyes_wont_2018,
+  title = {``{{Won}}'t {{Somebody Think}} of the {{Children}}?'' {{Examining COPPA Compliance}} at {{Scale}}},
+  shorttitle = {``{{Won}}'t {{Somebody Think}} of the {{Children}}?},
+  author = {Irwin Reyes and Primal Wijesekera and Joel Reardon and Amit Elazari Bar On and Abbas Razaghpanah and Narseo Vallina-Rodriguez and Serge Egelman},
+  year = {2018},
+  volume = {2018},
+  pages = {63--83},
+  issn = {2299-0984},
[…]
