+++ paper.bib
@@ -206,3 +206,13 @@
       publisher = ,
       month = aug,
 }
+
+@inbook{appchk,
+  author = {Geier, Oleg and Herrmann, Dominik},
+  year = {2021},
+  month = {06},
+  pages = {228-241},
+  title = {The AppChk Crowd-Sourcing Platform: Which Third Parties are iOS Apps Talking To?},
+  isbn = {978-3-030-78119-4},
+  doi = {10.1007/978-3-030-78120-0_15}
+}
