+++ README.md
@@ -218,28 +218,9 @@ Except where indicated otherwise, this project is licensed under [GPLv3](https:/
 
 ## Citation
 
-If you use this project as part of your academic studies, please kindly cite the below articles:
+If you use this project as part of your academic studies, please kindly cite the below article:
 
 ```
-@article{kollnig2022_iphone_android,
-      title={Are iPhones Really Better for Privacy? A Comparative Study of iOS and Android Apps}, 
-      author={Konrad Kollnig and Anastasia Shuba and Reuben Binns and Max {Van Kleek} and Nigel Shadbolt},
-      year={2022},
-      journal={Proceedings on Privacy Enhancing Technologies}
-}
-
-@article{kollnig_before_2021,
-      title = {{Before and after GDPR: Tracking in Mobile Apps}},
-      shorttitle = {{Before and after GDPR}},
-      author = {Kollnig, Konrad and Binns, Reuben and Van Kleek, Max and Lyngs, Ulrik and Zhao, Jun and Tinsman, Claudine and Shadbolt, Nigel},
-      year = {2021},
-      journaltitle = {Internet Policy Review},
-      volume = {10},
-      number = {4},
-      issn = {2197-6775},
-      doi = {10.14763/2021.4.1611}
-}
-
 @inproceedings {kollnig2021_consent,
       author = {Konrad Kollnig and Pierre Dewitte and Max Van Kleek and Ge Wang and Daniel Omeiza and Helena Webb and Nigel Shadbolt},
       title = {A Fait Accompli? An Empirical Study into the Absence of Consent to Third-Party Tracking in Android Apps},
