### 2022-04-04 #4270 @kasnder
A little update: I've now addressed most suggestions. The main missing piece is adding `javadocs` comments.

Additionally, I might add some screenshots for the now added example use, but then, this might not actually be necessary.
