### 2022-04-05 66b7639
Add JavaDoc and cleanup code
