### 2022-04-05 #4270 @kasnder
@gradvohl, thanks again for your swift review of our project and for taking out some of your time to do this. We've now tried to address your key points.

> ### About `paper.md` file
> Although the `paper.md` file is in the `joss` branch, it will be better if it is in the `master` branch, especially when the authors release the app's final version.

Good point! I've moved this to the `master` branch now.

> ### Mentions any ongoing research projects
> I did not find any mention about ongoing research projects. It is not a requirement for JOSS, but if the authors are aware of any project that uses the TrackControl app, I recommend citing that project or, perhaps, writing about its potential in other research projects.

Great suggestion! I've added a list of past and ongoing related research projects at the end of the `paper.md` file now.

> ### A list of key references
> There is a list of key references on the `paper.md` file. However, there are no references in the repository. I do not think that list is mandatory in the repository, but it is worth checking.

We have a paper that used TrackerControl as its methodology, and developed the underlying tracker database in a range of other publications.

I've added a list of key references to the repository now: https://github.com/TrackerControl/tracker-control-android#citation

> ### Example usage
> I did not find any tutorial or examples for using the software in any file on the repository. Therefore, I suggest that the author include a `docs` directory with a tutorial or screenshots describing the software's typical usage (step-by-step).

That's a great point, thanks! I've added detailed information on example use now: https://github.com/TrackerControl/tracker-control-android#example-use

> ### Scholarly effort
> The software had a substantial effort to build it. However, I consider that the software is much more commercial than academic software. I highlight that my opinion is not a demerit of the software.

Thanks for raising this. I think I see where this is coming from, and understand that our submission might be a little bit unusual. I think there are two aspects to the app: 1) the use by non-researchers, which is what drives the community and provides valuable feedback for app development, and 2) the use by researchers for legitimate research studies. On the latter, I've added a section to the `paper.md` file for clarification now.

> ### API Docs
> I did not find any documentation about the code regarding the API documentation, neither in the source code nor in a `docs` directory (the repository does not have a `docs` directory). Therefore, I strongly recommend the documentation of the source code using a standard. My suggestion is to use the Javadocs or [Doxygen](https://www.doxygen.nl/index.html) software. Doxygen can generate the docs (in HTML or PDF) from the documented source files.

Good point! I've added javadocs to the source code of the core of TrackerControl now (but not to the code that's been forked from NetGuard, since this is an external project and somewhat unrelated). I've used Doxygen to generate HTML in the `docs/` folder.

> ### Community guidelines
> There are indications of communities in different media (Telegram, Matrix etc). However, there are no community guidelines. My suggestion is to create a **Contributing** section (or a `CONTRIBUTING.md` file) with more specific instructions about how to contribute.

Another great point! I've now added a **Contributing** section: https://github.com/TrackerControl/tracker-control-android#contributing

> ### Tests on source code
> We ran the application tests on an Android smartphone. It works nicely. However, we did not find any automatic tests on the source code or instructions about how to do the tests. It is not mandatory to specify those instructions, but it would be nice if the authors provided instructions on how to compile and test the source code.

This is a good point! We're not doing tests because this is quite tricky in networked environments. I have, however, added build instructions now: https://github.com/TrackerControl/tracker-control-android#build-instructions

> ### Other Suggestions
> Please, consider including the `CITATION.cff` file with information about how to cite the `TrackerControl` software. In addition, information about software citation in GitHub is available on [GitHub Docs](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files).

Thanks, I wasn't even aware of this.. I've added such a file now: https://github.com/TrackerControl/tracker-control-android/blob/master/CITATION.cff

> ## Conclusion
> Analyzing all the aforementioned aspects, I rank the paper as **major revisions** needed. Although most of those revisions are very easy to implement, I consider the lack of documentation and the absence of the `paper.md` file in the master branch prevents acceptance of the paper at this time.

Please let us know if our changes are adequate, or if you think we should need to improve more.
