+++ CITATION.cff
@@ -8,5 +8,5 @@ authors:
   given-names: "Nigel"
   orcid: "https://orcid.org/0000-0002-5085-9724"
 title: "TrackerControl"
-date-released: 2017-12-18
+date-released: 2019-12-01
 url: "https://github.com/TrackerControl/tracker-control-android"
