### 2022-04-17 #4270 @gradvohl
I apologize for the delay in revisiting the submission. I am overwhelmed by the classes I am teaching, papers, and projects that I am writing (most of them with fast approaching deadlines).
Regarding the TrackerControl app, the authors addressed most of the issues I raised. However, there are some adjustments to those issues. I will point them as follows, using :heavy_check_mark: when it is ok; and :triangular_flag_on_post: when it needs adjustments.

### Mentions any ongoing research projects: :triangular_flag_on_post:
I am not sure if the authors addressed this issue properly. I did not find any mentions of ongoing projects at the end of the ``paper.md`` file. However, the paper mentions some works that served as a base for the TrackerControl app. Nevertheless, it is important to mention other projects that use the app or write about its potential in other research projects.

### A list of key references: :triangular_flag_on_post:
I did not find any key references list in the repository. The references are in the ``paper.md`` file, but not in the repository ``README.md`` file.

BTW, the ``README.md`` file has a ``Citation`` section. However, that section only shows how users should cite the TrackerControl app.

### Example Usage: :heavy_check_mark: (with observations)
It is ok. There is an example there. However, as a user, I would prefer to run the app other than reading all that text about how to use it. Therefore, I suggest that, in future releases, the authors add some screenshots showing typical software usage.

### API Docs: :heavy_check_mark: (with observations)
The authors generated the docs as HTML, but it does not look nice in the repository. If I am not wrong, Doxygen can generate a pdf file, which is more suitable for use, before cloning the repository. Therefore, please consider generating a pdf file in addition.

## Conclusion
Considering the revisions made by the authors, I believe the paper is closer to acceptance. Adjusting or justifying the points marked before with :triangular_flag_on_post: will lead to full acceptance, in my opinion.
