+++ paper.bib
@@ -38,6 +38,19 @@
   language = {en},
 }
 
+@inproceedings{le_antmonitor_2015,
+  title = {{{AntMonitor}}: {{A System}} for {{Monitoring}} from {{Mobile Devices}}},
+  shorttitle = {{{AntMonitor}}},
+  booktitle = {Proceedings of the 2015 {{ACM SIGCOMM Workshop}} on {{Crowdsourcing}} and {{Crowdsharing}} of {{Big}} ({{Internet}}) {{Data}} - {{C2B}}(1){{D}} '15},
+  author = {Le, Anh and Varmarken, Janus and Langhoff, Simon and Shuba, Anastasia and Gjoka, Minas and Markopoulou, Athina},
+  year = {2015},
+  pages = {15--20},
+  publisher = {{ACM Press}},
+  address = {{London, United Kingdom}},
+  doi = {10.1145/2787394.2787396},
+  isbn = {978-1-4503-3539-3}
+}
+
 @inproceedings{privacyguard_vpn_2015,
   author = {Yihang Song and Urs Hengartner},
   title = {{{PrivacyGuard}}: A {{VPN}}-based Platform to Detect Information Leakage on Android Devices},
@@ -70,6 +83,19 @@
   journal = {Proceedings on Privacy Enhancing Technologies},
   language = {en},
   number = {3},
+}
+
+@inproceedings{van_kleek_x-ray_2018,
+  title = {X-{{Ray Refine}}: {{Supporting}} the {{Exploration}} and {{Refinement}} of {{Information Exposure Resulting}} from {{Smartphone Apps}}},
+  shorttitle = {X-{{Ray Refine}}},
+  booktitle = {Proceedings of the 2018 {{CHI Conference}} on {{Human Factors}} in {{Computing Systems}}  - {{CHI}} '18},
+  author = {Van Kleek, Max and Binns, Reuben and Zhao, Jun and Slack, Adam and Lee, Sauyon and Ottewell, Dean and Shadbolt, Nigel},
+  year = {2018},
+  pages = {1--13},
+  publisher = {{ACM Press}},
+  address = {{Montreal QC, Canada}},
+  doi = {10.1145/3173574.3173967},
+  isbn = {978-1-4503-5620-6}
 }
 
 @InProceedings{van_kleek_better_2017,
@@ -188,11 +214,41 @@
   doi = {10.14722/ndss.2018.23009},
 }
 
+
+@article{kollnig_goodbye_2022,
+  title = {Goodbye {{Tracking}}? {{Impact}} of {{iOS App Tracking Transparency}} and {{Privacy Labels}}},
+  shorttitle = {Goodbye {{Tracking}}?},
+  author = {Kollnig, Konrad and Shuba, Anastasia and Van Kleek, Max and Binns, Reuben and Shadbolt, Nigel},
+  year = {2022},
+  journal = {arXiv:2204.03556 [cs]},
+  eprint = {2204.03556},
+  eprinttype = {arxiv},
+  primaryclass = {cs},
+  archiveprefix = {arXiv}
+}
+
 @article{kollnig2021_iphone_android,
-      title={Are iPhones Really Better for Privacy? A Comparative Study of iOS and Android Apps}, 
-      author={Konrad Kollnig and Anastasia Shuba and Reuben Binns and Max {Van Kleek} and Nigel Shadbolt},
-      year={2021},
-      journal={arXiv preprint arXiv:2109.13722}
+  author = {Konrad Kollnig and Anastasia Shuba and Reuben Binns and Max Van Kleek and Nigel Shadbolt},
+  doi = {10.2478/popets-2022-0033},
+  title = {Are {iPhones} Really Better for Privacy? A Comparative Study of {iOS} and {Android} Apps},
[…]
