+++ paper.md
@@ -32,19 +32,21 @@ In the past, the analysis of app tracking often used *off-device network analysi
 
 Other past studies have used *static analysis*, trying to gain insights into apps' data practices without executing them [@han_comparing_2013; @pios_2011]. Such static analysis has enabled the analysis of apps in the millions [@playdrone_2014; @wang_2018; @binns_third_2018], but usually does not to generate evidence of actual data transmission to trackers taking place, since apps are never run.
 
-In response, past research has developed on-device network analysis tools [@nomoads_2018; @privacyguard_vpn_2015; @lumen_2018]. Most of these used to operate at a domain-level and provided limited insights into what companies actually receive data from end-users, or to what countries data is sent and for what purpose. Furthermore, none of the previous on-device network analysis tools had been deployed at scale and most remained research prototypes.
+In response, past research has developed on-device network analysis tools [@le_antmonitor_2015; @nomoads_2018; @privacyguard_vpn_2015; @lumen_2018]. Most of these used to operate at a domain-level and provided limited insights into what companies actually receive data from end-users, or to what countries data is sent and for what purpose. Furthermore, none of the previous on-device network analysis tools had been deployed at scale and most remained research prototypes.
 
 # TrackerControl
 
 To improve the quality of tracking analysis and make it available to a wide audience, we developed the Android app TrackerControl (TC). This app provides users with real-time evidence of app tracking. TC analyses the network traffic of other apps, by establishing a local VPN on the Android phone and matching all observed network traffic against a database of known tracking domains. This allows to generate factual evidence of what companies apps share data with, and support research (both academic and non-academic) on app privacy.
 
-The tracking database behind TC is a unique feature of the app. The core of this database is the `X-Ray 2020` database that is the product of significant research efforts over the past years [@van_kleek_better_2017; @binns_third_2018;  @kollnig2021_consent; @kollnig2021_iphone_android]. This database has been created from analysing more than 2 million Android apps. The `X-Ray 2020` is complemented by the Disconnect.me database, that is the foundation for tracker blocking in Mozilla Firefox on the web. We further integrate the commonly used StevenBlack hostlist for tracking in apps, as a fallback. Overall, these databases provide information on 1) the *companies* behind tracking on the web and in apps, 2) the *countries* in which these companies are based, 3) and the *purposes* for which tracking is conducted (e.g. Analytics or Ads).
+The tracking database behind TC is a unique feature of the app. The core of this database is the `X-Ray 2020` database that is the product of significant research efforts over the past years [@van_kleek_better_2017; @binns_measuring_2018; @kollnig2021_consent; @kollnig2021_iphone_android]. This database has been created from analysing more than 2 million Android apps. The `X-Ray 2020` is complemented by the Disconnect.me database, that is the foundation for tracker blocking in Mozilla Firefox on the web. We further integrate the commonly used StevenBlack hostlist for tracking in apps, as a fallback. Overall, these databases provide information on 1) the *companies* behind tracking on the web and in apps, 2) the *countries* in which these companies are based, 3) and the *purposes* for which tracking is conducted (e.g. Analytics or Ads). The visualisation of tracking inside TC loosely follows the work by Van Kleek et al.&nbsp;[@van_kleek_better_2017; @van_kleek_x-ray_2018].
 
-The core of TC builds on the NetGuard app, which is in active use by millions of users worldwide [@netguard]. The high maturity of NetGuard ensures reliability of tracker analysis whilst minimising battery impact and supporting the long-term maintainability of TC.
[…]
