+++ paper.bib
@@ -169,6 +169,18 @@
   booktitle = {Proceedings of NDSS 2011},
 }
 
+@article{binns_measuring_2018,
+  title = {Measuring {{Third-party Tracker Power}} across {{Web}} and {{Mobile}}},
+  author = {Binns, Reuben and Zhao, Jun and Kleek, Max Van and Shadbolt, Nigel},
+  year = {2018},
+  journal = {ACM Transactions on Internet Technology},
+  volume = {18},
+  number = {4},
+  pages = {1--22},
+  issn = {15335399},
+  doi = {10.1145/3176246}
+}
+
 @inproceedings{wang_2018,
   author = {Haoyu Wang and Zhe Liu and Jingyue Liang and Narseo Vallina-Rodriguez and Yao Guo and Li Li and Juan Tapiador and Jingcun Cao and Guoai Xu},
   title = {Beyond {{Google Play}}: A Large-Scale Comparative Study of {{Chinese}} Android App Markets},
