+++ paper.bib
@@ -169,7 +169,7 @@
   booktitle = {Proceedings of NDSS 2011},
 }
 
-@article{binns_measuring_2018,
+@article{binns2018,
   title = {Measuring {{Third-party Tracker Power}} across {{Web}} and {{Mobile}}},
   author = {Binns, Reuben and Zhao, Jun and Kleek, Max Van and Shadbolt, Nigel},
   year = {2018},
@@ -177,7 +177,6 @@
   volume = {18},
   number = {4},
   pages = {1--22},
-  issn = {15335399},
   doi = {10.1145/3176246}
 }
 
