### 2022-04-23 #4270 @kasnder
Thank you for having another look at our work @gradvohl!

> ### About `paper.md` file ✔️
> ### Mentions any ongoing research projects: 🚩
> I am not sure if the authors addressed this issue properly. I did not find any mentions of ongoing projects at the end of the `paper.md` file. However, the paper mentions some works that served as a base for the TrackerControl app. Nevertheless, it is important to mention other projects that use the app or write about its potential in other research projects.

I added more details on this now, and a separate section in the `paper.md`.

> ### A list of key references: 🚩
> I did not find any key references list in the repository. The references are in the `paper.md` file, but not in the repository `README.md` file.
>
> BTW, the `README.md` file has a `Citation` section. However, that section only shows how users should cite the TrackerControl app.

I think I had misunderstood this. I've now added a list of key references to the `README.md` at <https://github.com/TrackerControl/tracker-control-android#references>.

> ### API Docs: ✔️ (with observations)
> The authors generated the docs as HTML, but it does not look nice in the repository. If I am not wrong, Doxygen can generate a pdf file, which is more suitable for use, before cloning the repository. Therefore, please consider generating a pdf file in addition.

I had some problems generating the pdf, but I've fixed this now and added the result at <https://github.com/TrackerControl/tracker-control-android/blob/master/docs/pdf/refman.pdf>.

> ## Conclusion
> Considering the revisions made by the authors, I believe the paper is closer to acceptance. Adjusting or justifying the points marked before with 🚩 will lead to full acceptance, in my opinion.

Do let us know if we've sufficiently addressed your suggestions, or if there's more work to do.
