### 2022-04-28 4a60ef0
Fix getColumnIndex lint errors
