+++ app/src/main/assets/ip_blocklist.txt
@@ -258,8 +258,6 @@
 12.14.172.204
 12.153.224.22
 12.158.234.169
-12.163.148.9
-12.163.148.99
 12.171.94.36
 12.171.94.38
 12.171.94.40
@@ -523,7 +521,6 @@
 162.144.0.192
 162.144.1.246
 162.42.211.167
-164.109.17.16
 164.115.2.134
 164.115.2.135
 164.115.2.144
@@ -1281,10 +1278,6 @@
 192.0.80.248
 192.104.182.105
 192.104.182.111
-192.114.68.46
-192.114.68.48
-192.114.68.50
-192.114.68.62
 192.148.26.68
 192.155.48.93
 192.155.50.93
@@ -2010,8 +2003,6 @@
 202.139.106.179
 202.139.107.20
 202.139.94.37
-202.168.212.103
-202.168.214.53
 202.170.120.110
 202.170.120.99
 202.172.236.75
@@ -2108,7 +2099,6 @@
 204.155.175.117
 204.155.175.201
 204.155.175.242
-204.16.173.2
 204.16.252.98
 204.160.120.125
 204.160.98.126
@@ -2171,7 +2161,6 @@
 205.150.0.180
 205.153.84.207
 205.153.85.31
-205.157.163.16
 205.177.28.48
 205.178.145.65
 205.178.152.39
@@ -2554,7 +2543,6 @@
 208.185.211.70
 208.185.211.71
 208.185.211.80
-208.185.247.210
 208.185.54.38
 208.185.54.44
 208.185.54.51
@@ -6247,7 +6235,6 @@
 64.237.101.92
 64.237.103.62
 64.237.108.209
-64.237.59.105
 64.237.61.69
 64.237.99.81
 64.239.183.120
@@ -6273,7 +6260,6 @@
 64.27.100.21
 64.27.100.214
 64.27.17.227
-64.27.17.233
 64.27.70.10
 64.27.90.240
 64.28.86.230
@@ -6898,8 +6884,6 @@
 66.230.192.41
 66.230.226.27
 66.230.239.174
-66.231.185.75
-66.231.185.81
 66.235.126.61
 66.235.126.71
 66.235.132.118
@@ -8407,7 +8391,6 @@
 78.46.67.131
 78.46.68.213
 78.46.69.205
-78.46.70.136
 78.46.9.18
 78.46.91.10
 78.47.133.245
@@ -8855,7 +8838,6 @@
 83.138.186.56
 83.138.187.17
 83.138.187.18
-83.138.67.159
 83.140.72.24
 83.141.36.24
 83.141.36.25
@@ -9121,13 +9103,6 @@
 88.198.215.209
 88.198.231.1
 88.198.231.8
-88.198.234.241
-88.198.234.242
-88.198.234.243
-88.198.234.244
-88.198.234.245
-88.198.234.246
-88.198.234.247
 88.198.24.51
 88.198.25.200
 88.198.34.76
@@ -9226,7 +9201,6 @@
 89.108.87.88
 89.108.87.90
 89.110.133.18
-89.110.148.165
 89.111.174.46
 89.144.1.148
 89.145.122.174
@@ -9651,7 +9625,6 @@
 94.127.74.60
 94.127.76.190
 94.127.76.60
-94.127.78.40
 94.136.40.82
 94.140.46.187
 94.198.100.232
