+++ app/src/main/assets/ip_blocklist.txt
@@ -904,7 +904,6 @@
 174.129.12.237
 174.129.14.235
 174.129.16.231
-174.129.17.142
 174.129.19.78
 174.129.2.142
 174.129.2.71
@@ -925,7 +924,6 @@
 174.129.232.188
 174.129.233.103
 174.129.233.188
-174.129.235.147
 174.129.236.152
 174.129.240.12
 174.129.243.62
@@ -933,10 +931,6 @@
 174.129.249.79
 174.129.252.108
 174.129.253.18
-174.129.28.138
-174.129.28.68
-174.129.29.15
-174.129.30.89
 174.129.35.49
 174.129.40.79
 174.129.43.138
@@ -1011,7 +1005,6 @@
 174.37.205.202
 174.37.207.102
 174.37.214.225
-174.37.234.56
 174.46.248.90
 174.47.188.40
 174.47.188.6
@@ -2085,7 +2078,6 @@
 204.13.248.125
 204.13.8.18
 204.13.8.76
-204.137.28.222
 204.137.28.46
 204.145.237.105
 204.145.71.140
@@ -3023,7 +3015,6 @@
 209.61.194.46
 209.62.105.12
 209.62.116.154
-209.62.126.2
 209.62.176.188
 209.62.176.73
 209.62.20.192
@@ -4312,7 +4303,6 @@
 216.201.83.161
 216.205.149.100
 216.205.150.13
-216.21.215.52
 216.21.222.7
 216.218.185.148
 216.218.211.56
@@ -7144,7 +7134,6 @@
 67.15.56.137
 67.15.58.29
 67.15.68.51
-67.15.68.54
 67.15.76.201
 67.15.77.192
 67.15.94.25
@@ -7728,7 +7717,6 @@
 72.2.7.31
 72.20.110.8
 72.20.114.94
-72.20.21.238
 72.20.40.27
 72.20.40.28
 72.20.40.29
@@ -8197,11 +8185,9 @@
 75.101.132.151
 75.101.135.45
 75.101.137.166
-75.101.137.205
 75.101.139.199
 75.101.140.131
 75.101.141.174
-75.101.141.193
 75.101.142.70
 75.101.143.85
 75.101.147.242
@@ -8209,7 +8195,6 @@
 75.101.153.39
 75.101.155.68
 75.101.158.158
-75.101.158.4
 75.101.159.219
 75.101.162.167
 75.101.163.223
@@ -9680,8 +9665,6 @@
 96.127.160.130
 96.30.19.32
 96.30.19.34
-96.30.8.104
-96.30.8.107
 96.31.88.157
 96.31.88.162
 96.44.154.35
