+++ fastlane/metadata/android/en-US/changelogs/2022052101.txt
@@ -1,0 +1,3 @@
+- Highlight tracker domains in traffic log as bold
+- Update host lists, dependencies, translations
+- Fix bugs
