### 2022-05-23 #4270 @sbenthall
Hello @gcdeshpande ,

This submission has gone through a round of reviewing by @gradvohl and revisions by the author.
But it is still waiting on a second review. Looking at [your checklist](https://github.com/openjournals/joss-reviews/issues/4270#issuecomment-1078715316), is it correct to say that you haven't reviewed the software yet?
Would it be possible for you to review the submission in its current status and see if it meets the requirements?

Thank you,
@sbenthall
