### 2022-05-26 26242ca
New Crowdin updates (#311)
* New translations strings.xml (Turkish)

* New translations strings.xml (Japanese)

* New translations full_description.txt (Japanese)
