+++ fastlane/metadata/android/ja-JP/full_description.txt
@@ -1,6 +1,6 @@
 TrackerControl を使用すると、普段気付いていない、広範囲にわたる、データ収集(トラッキング)を監視・制御できます。
 
-TrackerControl は、トラッキングを検出するため、Firefox で使用されている Disconnect blocklist と、 <i>2,000,000</i> 以上のアプリを解析することによって作成された、独自のブロックリストを組み合せて使用します！ さらに、カスタムブロックリストもサポートしています。
+TrackerControl は、Firefox で使用されている Disconnect blocklist と、 <i>2,000,000</i> 以上のアプリを解析することによって作成された独自のブロックリストを組み合せ、トラッキングを検出します！ さらに、カスタムブロックリストもサポートしています。
 
 できること
 - トラッキングの背後にある企業を明らかにし、
