### 2022-06-14 #4270 @sbenthall
@kasnder Could you please:
- make a tagged Git release for the final cut of the code
- submit an archive to e.g. Zenodo
- Report back the release version number and the archive DOI in this review thread.
