### 2022-06-15 807955f
Fix minor things in paper
