+++ app/src/main/assets/ip_blocklist.txt
@@ -422,7 +422,6 @@
 144.76.39.37
 146.101.66.152
 146.82.109.225
-146.82.138.194
 146.82.202.112
 146.82.202.121
 146.82.202.87
@@ -898,7 +897,6 @@
 174.123.55.226
 174.127.101.225
 174.127.127.178
-174.128.237.30
 174.129.102.210
 174.129.11.178
 174.129.12.237
@@ -1690,7 +1688,6 @@
 195.229.240.108
 195.235.97.80
 195.239.86.26
-195.24.37.245
 195.242.117.32
 195.245.125.226
 195.245.125.243
@@ -1962,14 +1959,12 @@
 200.147.3.170
 200.147.35.170
 200.147.67.170
-200.16.144.230
 200.173.74.162
 200.203.43.161
 200.221.7.120
 200.221.7.121
 200.221.7.122
 200.221.7.70
-200.32.3.53
 200.69.192.231
 200.69.192.239
 200.80.18.185
@@ -2187,7 +2182,6 @@
 205.209.188.53
 205.218.64.7
 205.218.65.142
-205.219.72.44
 205.234.175.175
 205.236.34.206
 205.236.34.7
@@ -2628,7 +2622,6 @@
 208.67.183.100
 208.67.195.100
 208.67.195.47
-208.67.28.20
 208.67.66.24
 208.67.70.27
 208.67.70.3
@@ -2858,7 +2851,6 @@
 209.126.137.252
 209.128.119.142
 209.128.76.107
-209.128.96.136
 209.132.205.101
 209.132.205.220
 209.132.205.221
@@ -4308,7 +4300,6 @@
 216.218.211.56
 216.218.211.57
 216.218.223.227
-216.218.255.25
 216.22.4.2
 216.22.4.5
 216.22.45.38
@@ -5121,7 +5112,6 @@
 4.78.243.32
 4.78.243.33
 4.78.243.34
-4.79.216.200
 42.120.219.6
 42.120.59.19
 42.120.59.20
@@ -6336,7 +6326,6 @@
 64.6.18.5
 64.6.28.5
 64.61.134.8
-64.62.171.123
 64.62.222.77
 64.62.232.32
 64.69.216.107
@@ -6650,8 +6639,6 @@
 65.75.162.56
 65.75.172.215
 65.98.53.169
-65.99.198.162
-65.99.198.180
 65.99.207.66
 65.99.253.170
 65.99.253.174
@@ -7437,7 +7424,6 @@
 69.43.161.132
 69.43.161.152
 69.43.161.162
-69.43.161.166
 69.43.161.177
 69.43.161.197
 69.43.161.210
@@ -7616,11 +7602,6 @@
 70.165.44.241
 70.182.179.196
 70.32.146.212
-70.39.84.249
-70.39.84.250
-70.39.99.86
-70.39.99.87
-70.39.99.88
 70.42.129.172
 70.42.129.173
 70.42.129.197
