+++ fastlane/metadata/android/en-US/changelogs/2022061501.txt
@@ -1,0 +1,1 @@
+- Update dependencies, translations, tracker lists
