### 2022-06-15 #4270 @kasnder
@sbenthall, here is the requested information:

Tagged release: <https://github.com/TrackerControl/tracker-control-android/releases/tag/2022061501>

Archive DOI: [10.5281/zenodo.6645297](https://doi.org/10.5281/zenodo.6645297)

Release version number: 2022061501
