### 2022-07-03 8da4298
New translations full_description.txt (Hebrew)
