### 2022-07-03 0336750
New translations short_description.txt (Hebrew)
