+++ fastlane/metadata/android/he-IL/short_description.txt
@@ -1,1 +1,1 @@
-TrackerControl allows to monitor and control hidden data collection in apps.
+TrackerControl מאפשר לנטר ולשלוט באיסוף נתונים מוסתר באפליקציות.
