### 2022-07-04 0bade78
New Crowdin updates (#320)
* New translations strings.xml (Persian)

* New translations full_description.txt (Persian)

* New translations short_description.txt (Persian)

* New translations strings.xml (Romanian)

* New translations strings.xml (Romanian)

* New translations strings.xml (Hebrew)

* New translations full_description.txt (Hebrew)

* New translations short_description.txt (Hebrew)

* New translations strings.xml (Hebrew)

* New translations strings.xml (Hebrew)

* New translations strings.xml (Hebrew)

* New translations full_description.txt (Hebrew)

* New translations short_de
[…]
