+++ fastlane/metadata/android/he-IL/short_description.txt
@@ -1,0 +1,1 @@
+TrackerControl מאפשר לנטר ולשלוט באיסוף נתונים מוסתר באפליקציות.
