+++ fastlane/metadata/android/he-IL/full_description.txt
@@ -1,0 +1,17 @@
+TrackerControl מאפשר למשתמשים לנטר ולשלוט באיסוף הנתונים הנרחב, המתמשך והנסתר באפליקציות לנייד על התנהגות המשתמש ('מעקב').
+
+כדי לזהות מעקב, TrackerControl משלב את העוצמה של רשימת החסימה של Disconnect, המשמשת את Firefox, ונעשה שימוש ברשימת החסימה הפנימית שלנו, שנוצרה מניתוח <i>~2,000,000</i> אפליקציות! בנוסף, TrackerControl תומך ברשימות חסימה מותאמות אישית.
+
+הגישה הזו
+- חושף את החברות מאחורי המעקב,
+- מאפשר לחסום מעקב באופן סלקטיבי, וכן
+חושף את מטרות המעקב, כגון ניתוח או פרסום.
+
+האפליקציה גם נועדה לחנך על הזכויות שלך על פי חוק הגנת מידע, כגון תקנת הגנת המידע הכללית של האיחוד האירופי (GDPR).
+
+מתחת למכסה המנוע, TrackerControl משתמש בפונקציונליות ה-VPN של אנדרואיד, כדי לנתח את תקשורת הרשת של אפליקציות <i>מקומית במכשיר האנדרואיד</i>. הדבר מתבצע באמצעות שרת VPN מקומי, כדי לאפשר ניתוח תעבורת רשת על ידי TrackerControl.
+
+אין צורך בשורש, VPNs אחרים או DNS פרטי אינם נתמכים. לא נעשה שימוש בשרת VPN חיצוני, כדי לשמור על בטיחות הנתונים שלך! TrackerControl אפילו מגן עליך מפני <i>הסואת DNS</i>, טכניקה פופולרית להסתרת עוקבים באתרים ובאפליקציות.
+
+TrackerControl תמיד יהיה חינמי וקוד פתוח, בהיותו פרויקט מחקר.
+
