### 2022-07-04 36d9681
Update and fix F-Droid page
