+++ fastlane/metadata/android/en-US/changelogs/2022070401.txt
@@ -1,0 +1,2 @@
+- Update translations (adds Hebrew)
+- Update dependencies
