### 2022-07-04 fdb4ac2
New translations full_description.txt (Hebrew)
