+++ fastlane/metadata/android/he-IL/full_description.txt
@@ -5,7 +5,7 @@ TrackerControl מאפשר למשתמשים לנטר ולשלוט באיסוף ה
 הגישה הזו
 - חושף את החברות מאחורי המעקב,
 - מאפשר לחסום מעקב באופן סלקטיבי, וכן
-חושף את מטרות המעקב, כגון ניתוח או פרסום.
+-חושף את מטרות המעקב, כגון ניתוח או פרסום.
 
 האפליקציה גם נועדה לחנך על הזכויות שלך על פי חוק הגנת מידע, כגון תקנת הגנת המידע הכללית של האיחוד האירופי (GDPR).
 
