+++ paper.bib
@@ -125,16 +125,6 @@
   language = {en},
 }
 
-@inproceedings{nomoads_2018,
-  author = {Anastasia Shuba and Athina Markopoulou and Zubair Shafiq},
-  year = {2018},
-  month = {10},
-  pages = {125-140},
-  title = {NoMoAds: Effective and Efficient Cross-App Mobile Ad-Blocking},
-  booktitle = {Proceedings on Privacy Enhancing Technologies 2018},
-  doi = {10.1515/popets-2018-0035},
-}
-
 @article{shuba_nomoats_2020,
   title = {{{NoMoATS}}: {{Towards Automatic Detection}} of {{Mobile Tracking}}},
   shorttitle = {{{NoMoATS}}},
@@ -225,17 +215,21 @@
   doi = {10.14722/ndss.2018.23009},
 }
 
-
-@article{kollnig_goodbye_2022,
-  title = {Goodbye {{Tracking}}? {{Impact}} of {{iOS App Tracking Transparency}} and {{Privacy Labels}}},
-  shorttitle = {Goodbye {{Tracking}}?},
-  author = {Kollnig, Konrad and Shuba, Anastasia and Van Kleek, Max and Binns, Reuben and Shadbolt, Nigel},
-  year = {2022},
-  journal = {arXiv:2204.03556 [cs]},
-  eprint = {2204.03556},
-  eprinttype = {arxiv},
-  primaryclass = {cs},
-  archiveprefix = {arXiv}
+@inproceedings{kollnig_goodbye_2022,
+	author = {Kollnig, Konrad and Shuba, Anastasia and Van Kleek, Max and Binns, Reuben and Shadbolt, Nigel},
+	title = {Goodbye Tracking? Impact of IOS App Tracking Transparency and Privacy Labels},
+	year = {2022},
+	isbn = {9781450393522},
+	publisher = {Association for Computing Machinery},
+	address = {New York, NY, USA},
+	url = {https://doi.org/10.1145/3531146.3533116},
+	doi = {10.1145/3531146.3533116},
+	booktitle = {2022 ACM Conference on Fairness, Accountability, and Transparency},
+	pages = {508–520},
+	numpages = {13},
+	keywords = {privacy, gatekeeper power, platform policies, mobile apps, Apple, App Tracking Transparency, data protection, iOS, Privacy Nutrition Labels},
+	location = {Seoul, Republic of Korea},
+	series = {FAccT '22}
 }
 
 @article{kollnig2021_iphone_android,
