### 2022-07-08 #4270 @kasnder
> @kasnder - I'm the AEiC this week, and I'll now proofread your paper, and then let you know if any changes are needed before publishing

Great, @danielskatz. Thanks for taking your time and being so swift!
