### 2022-07-08 e58ba05
paper/bib changes (#323)
* small changes in paper

* bib case fixes
