+++ paper.bib
@@ -39,7 +39,7 @@
 }
 
 @inproceedings{le_antmonitor_2015,
-  title = {{{AntMonitor}}: {{A System}} for {{Monitoring}} from {{Mobile Devices}}},
+  title = {{AntMonitor}: A System for Monitoring from Mobile Devices},
   shorttitle = {{{AntMonitor}}},
   booktitle = {Proceedings of the 2015 {{ACM SIGCOMM Workshop}} on {{Crowdsourcing}} and {{Crowdsharing}} of {{Big}} ({{Internet}}) {{Data}} - {{C2B}}(1){{D}} '15},
   author = {Le, Anh and Varmarken, Janus and Langhoff, Simon and Shuba, Anastasia and Gjoka, Minas and Markopoulou, Athina},
@@ -53,7 +53,7 @@
 
 @inproceedings{privacyguard_vpn_2015,
   author = {Yihang Song and Urs Hengartner},
-  title = {{{PrivacyGuard}}: A {{VPN}}-based Platform to Detect Information Leakage on Android Devices},
+  title = {{{PrivacyGuard}}: A {{VPN}}-based Platform to Detect Information Leakage on {A}ndroid Devices},
   booktitle = {Proceedings of the 5th Annual ACM CCS Workshop on Security and Privacy in Smartphones and Mobile Devices},
   series = {SPSM '15},
   year = {2015},
@@ -86,7 +86,7 @@
 }
 
 @inproceedings{van_kleek_x-ray_2018,
-  title = {X-{{Ray Refine}}: {{Supporting}} the {{Exploration}} and {{Refinement}} of {{Information Exposure Resulting}} from {{Smartphone Apps}}},
+  title = {{X-Ray Refine}: Supporting the Exploration and Refinement of Information Exposure Resulting from Smartphone Apps},
   shorttitle = {X-{{Ray Refine}}},
   booktitle = {Proceedings of the 2018 {{CHI Conference}} on {{Human Factors}} in {{Computing Systems}}  - {{CHI}} '18},
   author = {Van Kleek, Max and Binns, Reuben and Zhao, Jun and Slack, Adam and Lee, Sauyon and Ottewell, Dean and Shadbolt, Nigel},
@@ -99,7 +99,7 @@
 }
 
 @InProceedings{van_kleek_better_2017,
-  title = {Better the {{Devil You Know}}: {{Exposing}} the {{Data Sharing Practices}} of {{Smartphone Apps}}},
+  title = {Better the Devil You Know: Exposing the Data Sharing Practices of Smartphone Apps},
   shorttitle = {Better the {{Devil You Know}}},
   booktitle = {Proceedings of the 2017 {{CHI Conference}} on {{Human Factors}} in {{Computing Systems}} - {{CHI}} '17},
   author = {Max {Van Kleek} and Ilaria Liccardi and Reuben Binns and Jun Zhao and Daniel J. Weitzner and Nigel Shadbolt},
@@ -112,7 +112,7 @@
 }
 
 @InProceedings{ren_recon_2016,
-  title = {{{ReCon}}: {{Revealing}} and {{Controlling PII Leaks}} in {{Mobile Network Traffic}}},
+  title = {{ReCon}: Revealing and Controlling {PII} Leaks in Mobile Network Traffic},
   shorttitle = {{{ReCon}}},
   booktitle = {Proceedings of the 14th {{Annual International Conference}} on {{Mobile Systems}}, {{Applications}}, and {{Services}} - {{MobiSys}} '16},
   author = {Jingjing Ren and Ashwin Rao and Martina Lindorfer and Arnaud Legout and David Choffnes},
@@ -140,7 +140,7 @@
 }
 
 @inproceedings{han_comparing_2013,
[…]
