+++ paper.bib
@@ -172,7 +172,7 @@
 
 @inproceedings{wang_2018,
   author = {Haoyu Wang and Zhe Liu and Jingyue Liang and Narseo Vallina-Rodriguez and Yao Guo and Li Li and Juan Tapiador and Jingcun Cao and Guoai Xu},
-  title = {Beyond {{Google Play}}: A Large-Scale Comparative Study of {{Chinese}} Android App Markets},
+  title = {Beyond {{Google Play}}: A Large-Scale Comparative Study of {{Chinese}} {{Android}} App Markets},
   booktitle = {Proceedings of the Internet Measurement Conference 2018},
   series = {IMC '18},
   year = {2018},
